adiciona atualização de subtarefas já existentes e criação de novas
